Added initial commit of Twitch-RCON adapter.
This small utility allows channel point rewards on a Twitch channel to
trigger RCON commands on a Source engine game, Factorio, or Minecraft
server. This was intended as an indirect means to allow viewers to
trigger effects in Factorio via channel point rewards, but could also
work with Source engine games and Minecraft as well.