Added a config option for whether to retry RCON commands.
Previously, any RCON command that failed to execute due to a connection
error would be queued to be sent again after the RCON client reconnects.
However, in the event that a command causes the server to crash, this
could result in Twitch-RCON connecting to and immediately crashing the
server repeatedly.

The `retry` option defaults to `true`, to match the prior functionality.