Ensured that the RCON client will stop on Ctrl+C.
Previously, if the RCOn client is unable to connect and there are
messages queued up, the RCON client would continue attempting to
reconnect in order to send its messages before exiting.

Now, if it is connected it will still attempt te empty its queue, but it
will no longer attempt to reconnect after cancellation.